[rlwe]: added method to get galois elements for linear transform
-
